refactor(control-plane): split init and child-summary handlers
