+++ packages/control-plane/src/session/http/handlers/child-sessions.handler.ts
@@ -2,30 +2,16 @@ import { childFollowUpPromptRequestSchema } from "@open-inspect/shared/types/ses
 import { isSessionPromptable } from "@open-inspect/shared/types/session-activity";
 import { z } from "zod";
 import { sessionStatusSchema } from "@open-inspect/shared/types/sessions";
-import type { Logger } from "../../../logger";
 import { parsePersistedSandboxSettings } from "../../../sandbox/settings";
-import { parseArtifactMetadata } from "../../artifact-metadata";
 import type { SessionMessenger } from "../../messenger";
 import { PromptQueueFullError, SessionNotPromptableError } from "../../message-queue";
 import type { MessageRepository } from "../../message-repository";
-import type { ArtifactRepository } from "../../artifact-repository";
-import type { EventRepository } from "../../event-repository";
 import type { ParticipantRepository } from "../../participant-repository";
 import type { SessionCoreRepository } from "../../session-core-repository";
-import type { SandboxRepository } from "../../sandbox-repository";
 import type { MessageService } from "../../services/message.service";
 import type { SpawnContext } from "../../spawn-context";
 import { activePromptAuthorSchema, type ActivePromptAuthor } from "../../active-prompt-author";
-import { resolvePublicSessionId } from "../../public-session-id";
 import type { ParticipantRow } from "../../types";
-import {
-  RECENT_EVENT_FETCH_LIMIT,
-  buildChildSessionDetail,
-  collectFinalResponseEventRows,
-  parseChildSummaryOptions,
-  type ChildSummaryFinalResponseInput,
-  type ChildSummaryTrajectoryInput,
-} from "./child-session-summary";
 
 const parentPromptRequestSchema = childFollowUpPromptRequestSchema.extend({
   parentSessionId: z.string().min(1),
@@ -67,19 +53,16 @@ function toActivePromptAuthor(participant: ParticipantRow): ActivePromptAuthor {
 
 /**
  * HTTP boundary for the parent/child session endpoints: spawn context and
- * prompt-author reads for child spawning, the child summary read model, and
- * the parent-prompt/status-update callbacks children invoke.
+ * prompt-author reads for child spawning, and the parent-prompt/status-update
+ * callbacks children invoke. The child-summary read model lives in
+ * `ChildSummaryHandler` — its five collaborators are untouched by every
+ * route here.
  */
 export class ChildSessionsHandler {
   constructor(
     private readonly messageRepository: MessageRepository,
-    private readonly eventRepository: EventRepository,
     private readonly participantRepository: ParticipantRepository,
-    private readonly artifactRepository: ArtifactRepository,
     private readonly sessionCoreRepository: SessionCoreRepository,
-    private readonly sandboxRepository: SandboxRepository,
-    private readonly durableObjectId: string,
-    private readonly log: Logger,
     private readonly messenger: SessionMessenger,
     private readonly messageService: Pick<MessageService, "enqueuePrompt">
   ) {}
@@ -136,62 +119,6 @@ export class ChildSessionsHandler {
       this.participantRepository
     );
     return author instanceof Response ? author : Response.json(toActivePromptAuthor(author));
-  }
-
-  getChildSummary(url?: URL): Response {
-    const session = this.sessionCoreRepository.getSession();
-    if (!session) {
-      return Response.json({ error: "Session not found" }, { status: 404 });
-    }
-
-    const parsedOptions = parseChildSummaryOptions(url);
-    if (!parsedOptions.ok) {
-      return Response.json({ error: parsedOptions.error }, { status: 400 });
-    }
-
-    const options = parsedOptions.options;
-    const sandbox = this.sandboxRepository.getSandbox();
-    const artifacts = this.artifactRepository.listArtifacts();
-    const recentEventRows = this.eventRepository.listEventPage({
-      limit: RECENT_EVENT_FETCH_LIMIT,
-    }).events;
-    let finalResponse: ChildSummaryFinalResponseInput | undefined;
-    let trajectory: ChildSummaryTrajectoryInput | undefined;
-
-    if (options.includeFinalResponse) {
-      const terminalMessage = this.messageRepository.getLatestTerminalMessage();
-      const collectedEvents = terminalMessage
-        ? collectFinalResponseEventRows(this.eventRepository, terminalMessage.id)
-        : { eventRows: [], eventLimitReached: false };
-      finalResponse = { message: terminalMessage, ...collectedEvents };
-    }
-
-    if (options.includeTrajectory) {
-      const page = this.eventRepository.getEventTimelinePage({
-        limit: options.trajectoryLimit,
-        cursor: options.trajectoryCursor ?? undefined,
-      });
-      trajectory = {
-        eventRows: page.events,
-        hasMore: page.hasMore,
-        nextCursor: page.nextCursor,
-        limit: options.trajectoryLimit,
-      };
-    }
-
-    return Response.json(
-      buildChildSessionDetail({
-        session,
-        sandbox,
-        publicSessionId: resolvePublicSessionId(session, this.durableObjectId),
-        artifacts,
-        recentEventRows,
-        hasUnfinishedPrompt: this.messageRepository.getPendingOrProcessingCount() > 0,
-        parseArtifactMetadata: (artifact) => parseArtifactMetadata(artifact, this.log),
-        finalResponse,
-        trajectory,
-      })
-    );
   }
 
   async parentPrompt(request: Request): Promise<Response> {
